update Evaluation (attribute int->double)

+++ src/main/java/com/example/taximotoapp_backend/Evaluation/dto/request/EvaluationRequest.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class EvaluationRequest {
     private Long trajetId;
-    private int note;
+    private double note;
     private String commentaire;
 }
